Make contiv network installation files local
aci-gw and auth-proxy images, netplugin release, ovs rpm/deb files
will be loaded from local ansible files folders, if they exist.
If not we will continue to download from the urls or use docker
pull/docker run. This allows us to create a installation which is small
with external dependencies or larger but with no external dependencies
for contiv. Other roles are not modified as they are not part of the
core contiv installation.