Rewrite etcd role for proper lifecycle
Previously, docker could not be restarted on a node in a cluster and
have v2plugin remaining working due to a conflict because the container
was never stopped, the container name killed now matches that start name
for the container.

Previously, docker could not be shut down on multiple nodes without
getting etcd in a state where it could not start because the start/stop
script was dynamically managing the cluster membership as part of
service start/stop state.

Etcd is a HA database, starting or stopping should not depend on a
single node, the old code checked the first node for peers and used
that to determine if the cluster was being initialized or was already
active. Instead, just configure etcd statically with all the member
nodes in configuration.

To improve readability, much of the old etcd start/stop scripts
template to produce urls for etcd configuration was jinja macros,
redid that as role default variables using composition instead.

Introduced some naming for the etcd cluster that drives systemd service
naming

etcd member names are by default based on the hostname instead of
numbers that can be out of order with the machines.

The systemd unit was updated to handle:
* docker stop
* docker kill
* systemctl stop docker
* systemctl restart docker
When stopping docker, etcd will be stopped first, and when docker
starts, it will always trigger etcd to start.

Also, slow down restarts a little to avoid getting blocked by systemd
for restarting too many times and requiring manual intervention

v2plugin updated to be wantedby instead of PartOf so it starts when
docker is "start"ed and not just restarted

Drive-by:
* Add names to some plays and tasks to improve output
* Removed etcd peer port 7001 as it is not used

Signed-off-by: Chris Plock chrisplo@cisco.com